Move compute instance (and template) related helpers into their own file
